requirements.txt
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
requests==2.32.5
whitenoise==6.11.0